backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
flake8==5.0.4
flake8-black==0.3.3
Flake8-pyproject==1.1.0.post0
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytils==0.4.1
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
